refactor(app-tests): split oversized tech tree and map area test files
Reduce non-comment line count pressure for issue #2021 by splitting monolithic app test files into focused suites without changing test behavior, so line-limit enforcement can be adopted incrementally.

Made-with: Cursor
