+++ app/test/game_map_area_event_feed_test.dart
@@ -779,535 +779,4 @@ void main() {
       expect(lineFinder, findsOneWidget);
     },
   );
-
-  testWidgets('explore selection mode prompt appears under one second', (
-    WidgetTester tester,
-  ) async {
-    final init = getDebugInitGameResult();
-    final game = init.game;
-    final mapViewData = init.mapViewData;
-    final bus = AppEventBus.create();
-    addTearDown(bus.dispose);
-
-    final sampleUnitId = game.worldState.oldWorld.units.isNotEmpty
-        ? game.worldState.oldWorld.units.first.id
-        : game.worldState.newWorld.units.first.id;
-
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [
-          appEventBusProvider.overrideWith((ref) => bus),
-          currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-          gamesBoxProvider.overrideWith((ref) => gamesBox),
-          gameServiceProvider.overrideWith(
-            (ref) => GameService(gamesBox, GameSaveAdapter()),
-          ),
-          currentOrdersProvider.overrideWith(
-            () => CurrentOrdersNotifier(const Orders()),
-          ),
-          mapViewDataProvider.overrideWith((ref) => mapViewData),
-        ],
-        child: MaterialApp(
-          home: Scaffold(
-            body: GameMapArea(game: game, mapViewData: mapViewData),
-          ),
-        ),
-      ),
-    );
-    await tester.pump();
-    await tester.pump(const Duration(milliseconds: 16));
-
-    final sw = Stopwatch()..start();
-    bus.emit(
-      StartCivilianWorkTargetSelectionEvent(
-        unitId: sampleUnitId,
-        workTarget: kWorkTargetExplore,
-      ),
-    );
-    await tester.pump();
-
-    var selectionReady = false;
-    for (var i = 0; i < 200; i++) {
-      await tester.pump(const Duration(milliseconds: 5));
-      if (find.text('Select a tile, or click cancel').evaluate().isNotEmpty) {
-        selectionReady = true;
-        break;
-      }
-    }
-    sw.stop();
-
-    expect(selectionReady, isTrue);
-    expect(sw.elapsedMilliseconds, lessThan(1000));
-  });
-
-  testWidgets(
-    'build_improvement selection mode prompt appears under one second',
-    (WidgetTester tester) async {
-      final init = getDebugInitGameResult();
-      final game = init.game;
-      final mapViewData = init.mapViewData;
-      final bus = AppEventBus.create();
-      addTearDown(bus.dispose);
-
-      final humanPlayerId = game.players.firstWhere((p) => p.isHuman).id;
-      final topology = init.combinedTopology;
-      final playerView = buildPlayerView(game, topology, humanPlayerId);
-
-      String? builderUnitId;
-      for (final unit in [
-        ...game.worldState.oldWorld.units,
-        ...game.worldState.newWorld.units,
-      ]) {
-        if (!(workOrderTargetsByUnitType[unit.type]?.contains(
-              kWorkTargetBuildImprovement,
-            ) ??
-            false)) {
-          continue;
-        }
-        final valid = getValidWorkOrderTileKeysWithVisibility(
-          game: game,
-          topology: topology,
-          view: playerView,
-          unitId: unit.id,
-          workTarget: kWorkTargetBuildImprovement,
-          currentOrders: const Orders(),
-          tileMapByRegion: init.tileMapByRegion,
-        );
-        if (valid.isNotEmpty) {
-          builderUnitId = unit.id;
-          break;
-        }
-      }
-
-      expect(
-        builderUnitId,
-        isNotNull,
-        reason:
-            'debug init fixture must include a builder with at least one valid '
-            'build_improvement target tile',
-      );
-
-      await tester.pumpWidget(
-        ProviderScope(
-          overrides: [
-            appEventBusProvider.overrideWith((ref) => bus),
-            currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-            gamesBoxProvider.overrideWith((ref) => gamesBox),
-            gameServiceProvider.overrideWith(
-              (ref) => GameService(gamesBox, GameSaveAdapter()),
-            ),
-            currentOrdersProvider.overrideWith(
-              () => CurrentOrdersNotifier(const Orders()),
-            ),
-            mapViewDataProvider.overrideWith((ref) => mapViewData),
-          ],
-          child: MaterialApp(
-            home: Scaffold(
-              body: GameMapArea(game: game, mapViewData: mapViewData),
-            ),
-          ),
-        ),
-      );
-      await tester.pump();
-      await tester.pump(const Duration(milliseconds: 16));
-
-      final sw = Stopwatch()..start();
-      bus.emit(
-        StartCivilianWorkTargetSelectionEvent(
-          unitId: builderUnitId!,
-          workTarget: kWorkTargetBuildImprovement,
-        ),
-      );
-      await tester.pump();
-
-      var selectionReady = false;
-      for (var i = 0; i < 200; i++) {
-        await tester.pump(const Duration(milliseconds: 5));
-        if (find.text('Select a tile, or click cancel').evaluate().isNotEmpty) {
-          selectionReady = true;
-          break;
-        }
-      }
-      sw.stop();
-
-      expect(selectionReady, isTrue);
-      expect(sw.elapsedMilliseconds, lessThan(1000));
-    },
-  );
-
-  testWidgets(
-    'work target selection auto-switches to region with valid tiles when current tab has none',
-    (WidgetTester tester) async {
-      final init = getDebugInitGameResult();
-      final game = init.game;
-      final mapViewData = init.mapViewData;
-      final bus = AppEventBus.create();
-      addTearDown(bus.dispose);
-
-      final humanPlayerId = game.players.firstWhere((p) => p.isHuman).id;
-      final workTargets = <String>[
-        kWorkTargetExplore,
-        kWorkTargetProspect,
-        kWorkTargetBuildImprovement,
-        kWorkTargetUpgradeTown,
-        kWorkTargetBuildRoad,
-        kWorkTargetBuildPort,
-        kWorkTargetBuildFort,
-        kWorkTargetBuildRail,
-        kWorkTargetStealTech,
-        kWorkTargetCounterSpy,
-        kWorkTargetPurchaseLand,
-      ];
-      final topology = init.combinedTopology;
-      final playerView = buildPlayerView(game, topology, humanPlayerId);
-      final unitById = <String, Unit>{
-        for (final unit in [
-          ...game.worldState.oldWorld.units,
-          ...game.worldState.newWorld.units,
-        ])
-          unit.id: unit,
-      };
-
-      ({String unitId, String workTarget})? offTabSelection;
-      for (final unit in unitById.values) {
-        for (final workTarget in workTargets) {
-          final valid = getValidWorkOrderTileKeysWithVisibility(
-            game: game,
-            topology: topology,
-            view: playerView,
-            unitId: unit.id,
-            workTarget: workTarget,
-            currentOrders: const Orders(),
-            tileMapByRegion: init.tileMapByRegion,
-          );
-          final hasOldWorld = valid.any((k) => k.startsWith('oldWorld|'));
-          final hasOnlyOldWorld =
-              hasOldWorld && !valid.any((k) => k.startsWith('newWorld|'));
-          if (hasOnlyOldWorld) {
-            offTabSelection = (unitId: unit.id, workTarget: workTarget);
-            break;
-          }
-        }
-        if (offTabSelection != null) {
-          break;
-        }
-      }
-
-      expect(
-        offTabSelection,
-        isNotNull,
-        reason:
-            'debug init fixture must include one selection with valid tiles only in Old World',
-      );
-
-      await tester.pumpWidget(
-        ProviderScope(
-          overrides: [
-            appEventBusProvider.overrideWith((ref) => bus),
-            currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-            gamesBoxProvider.overrideWith((ref) => gamesBox),
-            gameServiceProvider.overrideWith(
-              (ref) => GameService(gamesBox, GameSaveAdapter()),
-            ),
-            currentOrdersProvider.overrideWith(
-              () => CurrentOrdersNotifier(const Orders()),
-            ),
-            mapViewDataProvider.overrideWith((ref) => mapViewData),
-          ],
-          child: MaterialApp(
-            home: Scaffold(
-              body: GameMapArea(game: game, mapViewData: mapViewData),
-            ),
-          ),
-        ),
-      );
-      await tester.pump();
-      await tester.pump(const Duration(milliseconds: 16));
-
-      await tester.tap(find.text('New World'));
-      await tester.pump();
-      await tester.pump(const Duration(milliseconds: 16));
-
-      bus.emit(
-        StartCivilianWorkTargetSelectionEvent(
-          unitId: offTabSelection!.unitId,
-          workTarget: offTabSelection.workTarget,
-        ),
-      );
-      await tester.pump();
-      await tester.pump(const Duration(milliseconds: 16));
-
-      final autoSwitchedRegionMap = tester
-          .widgetList<CtRegionMap>(find.byType(CtRegionMap))
-          .first;
-      final autoSwitchedValidKeys = autoSwitchedRegionMap.validTileKeys;
-      expect(autoSwitchedRegionMap.region.regionId, kRegionOldWorld);
-      expect(autoSwitchedValidKeys, isNotNull);
-      expect(
-        autoSwitchedValidKeys!.every((k) => k.startsWith('oldWorld|')),
-        isTrue,
-      );
-
-      final afterSwitchRegionMap = tester
-          .widgetList<CtRegionMap>(find.byType(CtRegionMap))
-          .first;
-      final afterSwitchValidKeys = afterSwitchRegionMap.validTileKeys;
-      expect(afterSwitchValidKeys, isNotNull);
-      expect(
-        afterSwitchValidKeys!.every((k) => k.startsWith('oldWorld|')),
-        isTrue,
-      );
-    },
-  );
-
-  testWidgets(
-    'work target selection shows prompt overlay and cancel button exits mode',
-    (WidgetTester tester) async {
-      final init = getDebugInitGameResult();
-      final game = init.game;
-      final mapViewData = init.mapViewData;
-      final bus = AppEventBus.create();
-      addTearDown(bus.dispose);
-
-      final sampleUnitId = game.worldState.oldWorld.units.isNotEmpty
-          ? game.worldState.oldWorld.units.first.id
-          : game.worldState.newWorld.units.first.id;
-
-      await tester.pumpWidget(
-        ProviderScope(
-          overrides: [
-            appEventBusProvider.overrideWith((ref) => bus),
-            currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-            gamesBoxProvider.overrideWith((ref) => gamesBox),
-            gameServiceProvider.overrideWith(
-              (ref) => GameService(gamesBox, GameSaveAdapter()),
-            ),
-            currentOrdersProvider.overrideWith(
-              () => CurrentOrdersNotifier(const Orders()),
-            ),
-            mapViewDataProvider.overrideWith((ref) => mapViewData),
-          ],
-          child: MaterialApp(
-            home: Scaffold(
-              body: GameMapArea(game: game, mapViewData: mapViewData),
-            ),
-          ),
-        ),
-      );
-      await tester.pump();
-      await tester.pump(const Duration(milliseconds: 16));
-
-      bus.emit(
-        StartCivilianWorkTargetSelectionEvent(
-          unitId: sampleUnitId,
-          workTarget: kWorkTargetExplore,
-        ),
-      );
-      await tester.pump();
-      await tester.pump(const Duration(milliseconds: 16));
-
-      expect(find.text('Select a tile, or click cancel'), findsOneWidget);
-      expect(find.text('cancel'), findsOneWidget);
-
-      await tester.tap(find.text('cancel'));
-      await tester.pump();
-
-      expect(find.text('Select a tile, or click cancel'), findsNothing);
-    },
-  );
-
-  testWidgets('left rail icon cancels selection mode before opening panel', (
-    WidgetTester tester,
-  ) async {
-    final init = getDebugInitGameResult();
-    final game = init.game;
-    final mapViewData = init.mapViewData;
-    final bus = AppEventBus.create();
-    addTearDown(bus.dispose);
-
-    final openedPanels = <OpenCivilianUnitsPanelEvent>[];
-    final panelSub = bus.on<OpenCivilianUnitsPanelEvent>().listen(
-      openedPanels.add,
-    );
-    addTearDown(panelSub.cancel);
-
-    final sampleUnitId = game.worldState.oldWorld.units.isNotEmpty
-        ? game.worldState.oldWorld.units.first.id
-        : game.worldState.newWorld.units.first.id;
-
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [
-          appEventBusProvider.overrideWith((ref) => bus),
-          currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-          gamesBoxProvider.overrideWith((ref) => gamesBox),
-          gameServiceProvider.overrideWith(
-            (ref) => GameService(gamesBox, GameSaveAdapter()),
-          ),
-          currentOrdersProvider.overrideWith(
-            () => CurrentOrdersNotifier(const Orders()),
-          ),
-          mapViewDataProvider.overrideWith((ref) => mapViewData),
-        ],
-        child: MaterialApp(
-          home: Scaffold(
-            body: GameMapArea(game: game, mapViewData: mapViewData),
-          ),
-        ),
-      ),
-    );
-    await tester.pump();
-    await tester.pump(const Duration(milliseconds: 16));
-
-    bus.emit(
-      StartCivilianWorkTargetSelectionEvent(
-        unitId: sampleUnitId,
-        workTarget: kWorkTargetExplore,
-      ),
-    );
-    await tester.pump();
-    await tester.pump(const Duration(milliseconds: 16));
-    expect(find.text('Select a tile, or click cancel'), findsOneWidget);
-
-    await tester.tap(find.byKey(kEmpireCivilianUnitsButtonKey));
-    await tester.pump();
-
-    expect(find.text('Select a tile, or click cancel'), findsNothing);
-    expect(openedPanels, hasLength(1));
-  });
-
-  testWidgets('escape key cancels work target selection mode', (
-    WidgetTester tester,
-  ) async {
-    final init = getDebugInitGameResult();
-    final game = init.game;
-    final mapViewData = init.mapViewData;
-    final bus = AppEventBus.create();
-    addTearDown(bus.dispose);
-
-    final sampleUnitId = game.worldState.oldWorld.units.isNotEmpty
-        ? game.worldState.oldWorld.units.first.id
-        : game.worldState.newWorld.units.first.id;
-
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [
-          appEventBusProvider.overrideWith((ref) => bus),
-          currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-          gamesBoxProvider.overrideWith((ref) => gamesBox),
-          gameServiceProvider.overrideWith(
-            (ref) => GameService(gamesBox, GameSaveAdapter()),
-          ),
-          currentOrdersProvider.overrideWith(
-            () => CurrentOrdersNotifier(const Orders()),
-          ),
-          mapViewDataProvider.overrideWith((ref) => mapViewData),
-        ],
-        child: MaterialApp(
-          home: Scaffold(
-            body: GameMapArea(game: game, mapViewData: mapViewData),
-          ),
-        ),
-      ),
-    );
-    await tester.pump();
-    await tester.pump(const Duration(milliseconds: 16));
-
-    bus.emit(
-      StartCivilianWorkTargetSelectionEvent(
-        unitId: sampleUnitId,
-        workTarget: kWorkTargetExplore,
-      ),
-    );
-    await tester.pump();
-    await tester.pump(const Duration(milliseconds: 16));
-
-    expect(find.text('Select a tile, or click cancel'), findsOneWidget);
-    var regionMap = tester.widget<CtRegionMap>(find.byType(CtRegionMap).first);
-    expect(regionMap.validTileKeys, isNotNull);
-
-    final keyHandlerFocuses = tester
-        .widgetList<Focus>(find.byType(Focus))
-        .where((focus) => focus.onKeyEvent != null);
-    var keyHandled = false;
-    for (final focus in keyHandlerFocuses) {
-      final keyResult = focus.onKeyEvent!(
-        FocusNode(),
-        const KeyDownEvent(
-          timeStamp: Duration.zero,
-          logicalKey: LogicalKeyboardKey.escape,
-          physicalKey: PhysicalKeyboardKey.escape,
-        ),
-      );
-      if (keyResult == KeyEventResult.handled) {
-        keyHandled = true;
-        break;
-      }
-    }
-    expect(keyHandled, isTrue);
-    await tester.pump();
-    await tester.pump(const Duration(milliseconds: 16));
-
-    expect(find.text('Select a tile, or click cancel'), findsNothing);
-    regionMap = tester.widget<CtRegionMap>(find.byType(CtRegionMap).first);
-    expect(regionMap.validTileKeys, isNull);
-  });
-
-  testWidgets('selection mode blocks non-selection map interaction callbacks', (
-    WidgetTester tester,
-  ) async {
-    final init = getDebugInitGameResult();
-    final game = init.game;
-    final mapViewData = init.mapViewData;
-    final bus = AppEventBus.create();
-    addTearDown(bus.dispose);
-
-    final sampleUnitId = game.worldState.oldWorld.units.isNotEmpty
-        ? game.worldState.oldWorld.units.first.id
-        : game.worldState.newWorld.units.first.id;
-
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [
-          appEventBusProvider.overrideWith((ref) => bus),
-          currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-          gamesBoxProvider.overrideWith((ref) => gamesBox),
-          gameServiceProvider.overrideWith(
-            (ref) => GameService(gamesBox, GameSaveAdapter()),
-          ),
-          currentOrdersProvider.overrideWith(
-            () => CurrentOrdersNotifier(const Orders()),
-          ),
-          mapViewDataProvider.overrideWith((ref) => mapViewData),
-        ],
-        child: MaterialApp(
-          home: Scaffold(
-            body: GameMapArea(game: game, mapViewData: mapViewData),
-          ),
-        ),
-      ),
-    );
-    await tester.pump();
-    await tester.pump(const Duration(milliseconds: 16));
-
-    var regionMap = tester.widget<CtRegionMap>(find.byType(CtRegionMap).first);
-    expect(regionMap.onMapTileTappedForDetail, isNotNull);
-    expect(regionMap.onCivilianTileTapped, isNotNull);
-    expect(regionMap.onFleetMarkerTapped, isNotNull);
-
-    bus.emit(
-      StartCivilianWorkTargetSelectionEvent(
-        unitId: sampleUnitId,
-        workTarget: kWorkTargetExplore,
-      ),
-    );
-    await tester.pump();
-    await tester.pump(const Duration(milliseconds: 16));
-
-    expect(find.text('Select a tile, or click cancel'), findsOneWidget);
-    regionMap = tester.widget<CtRegionMap>(find.byType(CtRegionMap).first);
-    expect(regionMap.onMapTileTappedForDetail, isNull);
-    expect(regionMap.onCivilianTileTapped, isNull);
-    expect(regionMap.onFleetMarkerTapped, isNull);
-  });
 }
